@@ -2,7 +2,7 @@ apply plugin: 'com.android.application'
 apply plugin: 'com.onesignal.androidsdk.onesignal-gradle-plugin'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 29
     defaultConfig {
         manifestPlaceholders = [
                 onesignal_app_id: '0626983b-4f76-47c4-bc41-01321b89e700',
@@ -11,10 +11,10 @@ android {
         ]
         applicationId "com.sattar.j.smsfake"
         minSdkVersion 17
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 9
         versionName "1.0.9.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
     lintOptions {
         abortOnError false
@@ -29,10 +29,10 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support:cardview-v7:27.1.1'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.2'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     testImplementation 'junit:junit:4.12'
     implementation 'com.github.marcoscgdev:DialogSheet:1.0.5'
     implementation 'com.jaredrummler:material-spinner:1.2.5'
@@ -50,4 +50,6 @@ dependencies {
     //**
     implementation 'com.infideap.drawerbehavior:drawer-behavior:0.0.5'
 
+    implementation 'com.github.bumptech.glide:glide:4.11.0'
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.11.0'
 }
