@@ -14,15 +14,15 @@ import android.os.Handler;
 import android.preference.PreferenceManager;
 import android.provider.ContactsContract;
 import android.provider.Telephony;
-import android.support.annotation.NonNull;
-import android.support.annotation.RequiresApi;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.NavigationView;
-import android.support.v4.view.GravityCompat;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.telephony.gsm.SmsManager;
+import androidx.annotation.NonNull;
+import androidx.annotation.RequiresApi;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.google.android.material.navigation.NavigationView;
+import androidx.core.view.GravityCompat;
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+
 import android.util.Log;
 import android.view.Gravity;
 import android.view.LayoutInflater;
@@ -37,6 +37,8 @@ import com.github.florent37.materialtextfield.MaterialTextField;
 import com.infideap.drawerbehavior.Advance3DDrawerLayout;
 import com.jaredrummler.materialspinner.MaterialSpinner;
 import com.marcoscg.dialogsheet.DialogSheet;
+import com.sattar.j.smsfake.CustomViews.CustomEditText;
+import com.sattar.j.smsfake.CustomViews.CustomTextView;
 
 import cn.pedant.SweetAlert.SweetAlertDialog;
 import co.ronash.pushe.Pushe;
@@ -48,8 +50,8 @@ public class MainActivity extends AppCompatActivity implements NavigationView.On
     private MaterialTextField phone;
     private FloatingActionButton send;
     private String sender, body, result = "0";
-    private MyEditText senderInput, bodyInput;
-    private MyTextView timeTxt, dateTxt;
+    private CustomEditText senderInput, bodyInput;
+    private CustomTextView timeTxt, dateTxt;
     private MaterialSpinner spinner;
     private SweetAlertDialog pDialog;
     private ContentValues values;
